requirements.txt
django
django-heroku
django-bootstrap4
django-bootstrap-datepicker-plus
django-cleanup
django-widget-tweaks
dj_database_url
djangorestframework
geopy
gunicorn
htmldom
Pillow
pytz
requests
